+++ src/pymeshlab/bindings/pymesh.cpp
@@ -79,6 +79,7 @@ void pymeshlab::initMesh(py::module& m)
 	meshClass.def("compact", &Mesh::compact, doc::PYMESH_COMPACT);
 
 	meshClass.def("vertex_matrix", &meshlab::vertexMatrix, doc::PYMESH_VERT_MAT);
+	meshClass.def("edge_matrix", &meshlab::edgeMatrix, doc::PYMESH_EDGE_MAT);
 	meshClass.def("face_matrix", &meshlab::faceMatrix, doc::PYMESH_FACE_MAT);
 	meshClass.def("polygonal_face_list", &meshlab::polygonalFaceList, doc::PYMESH_POLYGONAL_FACE_LIST);
 	meshClass.def("vertex_normal_matrix", &meshlab::vertexNormalMatrix, doc::PYMESH_VERT_NORM_MAT);